refactor de MediaBox a Media, de layout a feature
